backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-chroma
pypdf
chromadb
google-generativeai
